requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
asgiref==3.5.0
async-timeout==4.0.2
attrs==21.4.0
backports.zoneinfo==0.2.1
base58==2.1.1
bcrypt==3.1.7
bitarray==1.2.2
blinker==1.4
certifi==2019.11.28
chardet==3.0.4
charset-normalizer==2.0.12
Click==7.0
colorama==0.4.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.8
cytoolz==0.11.2
dj-database-url==0.5.0
Django==4.0.3
django-cors-headers==3.11.0
djangorestframework==3.13.1
drf-yasg==1.20.0
entrypoints==0.3
eth-abi==2.1.1
eth-account==0.5.7
eth-hash==0.3.2
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.10.0
fasteners==0.14.1
frozenlist==1.3.0
future==0.18.2
hexbytes==0.2.2
httplib2==0.14.0
idna==2.8
inflection==0.5.1
ipfshttpclient==0.8.0a2
itypes==1.2.0
Jinja2==3.1.1
jsonschema==3.2.0
keyring==18.0.1
gunicorn==20.1.0
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
lru-dict==1.1.7
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==2.1.1
monotonic==1.5
multiaddr==0.0.9
multidict==6.0.2
netaddr==0.8.0
netifaces==0.10.4
numpy==1.22.3
oauthlib==3.1.0
olefile==0.46
packaging==21.3
pandas==1.4.2
pandas-ta==0.3.14b0
paramiko==2.6.0
parsimonious==0.8.1
pexpect==4.6.0
Pillow==7.0.0
psycopg2-binary
protobuf==3.20.0
pycairo==1.16.2
pycryptodome==3.14.1
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==3.0.8
pyRFC3339==1.1
pyrsistent==0.18.1
python-dateutil
pytz
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
rlp==2.0.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
SecretStorage==2.3.1
simplejson==3.16.0
six==1.14.0
sqlparse==0.4.2
systemd-python==234
toolz==0.11.2
uritemplate==4.1.1
urllib3==1.25.8
varint==1.0.2
wadllib==1.3.3
web3==5.28.0
whitenoise==6.0.0
websockets==9.1
yarl==1.7.2
